requirements.txt
Django>=4.2,<6.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.0.0
django-walletpass>=0.1.0
cryptography>=41.0.0
Pillow>=10.0.0
requests>=2.31.0
PyJWT>=2.0.0
h2>=4.0.0
django-cors-headers>=4.0.0
django-storages>=1.14.0